feat(altitude): implement automated altitude reference discovery
* Refactor KnownLocationsDatabaseManager to support weighted moving average altitude learning.
* Upgrade StartLocation2Altitude.db to Version 4, adding hitCount and double-precision support.
* Integrate background learning hook into WorkoutRepository to automatically discover and refine reference points upon workout completion.
* Refactor AltitudeFromPressureDevice to utilize high-precision discovered altitudes for barometric calibration.
* Eliminate the need for manual user maintenance of the start locations database.
* Fulfills REQ-DAT-007; resolves ATT-39.

+++ app/src/main/java/com/atrainingtracker/banalservice/devices/AltitudeFromPressureDevice.java
@@ -128,7 +128,7 @@ public class AltitudeFromPressureDevice extends MyDevice
     /**
      * set the field mAltitudeCorrection
      */
-    private void setAltitudeCorrection(int correctAltitude) {
+    private void setAltitudeCorrection(double correctAltitude) {
         if (DEBUG) Log.d(TAG, "setAltitudeCorrection");
 
         mAltitudeCorrection = correctAltitude - mAltitudeSensor.getValue().doubleValue();
